// src/app/modules/landing/order-details/order-details.type.ts
export interface Order
{
    appliedDiscount: number;
    appliedDiscountDescription: string;
    beingProcess: string;
    cartId: string;
    completionStatus: string;
    created: string;
    customer: string;
    customerId: string;
    customerNotes: string;
    deliveryCharges: number;
    deliveryDiscount: number;
    deliveryDiscountDescription: string;
    deliveryDiscountMaxAmount: number;
    deliveryType: string;
    discountCalculationType: string;
    discountCalculationValue: number;
    discountId: string;
    discountMaxAmount: number;
    id: string;
    invoiceId: string;
    klCommission: number;
    orderPaymentDetail: {
        accountName: string;
        couponId: string;
        deliveryQuotationAmount:  number;
        deliveryQuotationReferenceId: string;
        gatewayId: string;
        orderId: string;
        time: string;
    }
    orderShipmentDetail: {
        address: string;
        city: string;
        country: string;
        customerTrackingUrl: string;
        deliveryProviderId: string;
        deliveryType: string;
        email: string;
        merchantTrackingUrl: string;
        orderId: string;
        phoneNumber: string;
        receiverName: string;
        state: string;
        storePickup: boolean;
        trackingNumber: string;
        trackingUrl: string;
        zipcode: string;
    }
    paymentStatus: string;
    paymentType: string;
    privateAdminNotes: string;
    store: {
        id: string,
        name: string,
        city: string,
        address: string,
        clientId: string,
        postcode: string,
        state: string,
        contactName: string,
        phone: string,
        phoneNumber: string,
        email: string,
        verticalCode: string,
        serviceChargesPercentage: number,
        paymentType: string,
        invoiceSeqNo: number,
        nameAbreviation: string
    }
    storeId: string;
    storeServiceCharges: number;
    storeShare: number;
    subTotal: number;
    total: number;
    updated: string;
}